Don't overuse exceptions
We were using exceptions to mark "normal" failure cases like the
user provided a too-short string. This was fragile: before, we raised a
ShortLocationError when a query was less than 3 letters long.
But, in the actual app routes,
we forgot to include this in the list of exceptions we caught.
Thus, curently in prod if
you ask for "XX" or similar the you get as a response:
`{"Code":"InternalServerError","Message":"An internal server error occurred."`
which doesn't show up as anything via twilio or on the
website.

This is good evidence that throwing errors for normal failures is bad.
We can jsut as easily just return
a simple error string.

There IS a case where the exception is useful, which is where we call
`geocode()` and we don;t want it to return EITHER a dict or None, we
just want it to return a dict. So keep doing that.

This also adds two new log statements to differentiate different kinds
of journeys.